+++ spec/dummy/config/webpack/webpack.config.js
@@ -7,10 +7,10 @@ const envSpecificConfig = () => {
   const path = resolve(__dirname, `${env.nodeEnv}.js`);
   if (existsSync(path)) {
     console.log(`Loading ENV specific webpack configuration file ${path}`);
+    // eslint-disable-next-line import/no-dynamic-require, global-require
     return require(path);
-  } else {
-    return generateWebpackConfig();
   }
+  return generateWebpackConfig();
 };
 
 module.exports = envSpecificConfig();